Pass the struct!!
Passing the packet struct is the way that I want to do it.
For some reson, it keeps getting passed up as empty.
Not sure why. I think it has to do with memory access between
processes?

+++ mp/star_packet.c
@@ -80,7 +80,7 @@ char* packetString(star_packet* packet) {
 
 int findSecondAsterisk(char* packetString) {
    int i, seen = 0;
-   for (i = 0; i < strlen(packetString); i++) {
+   for (i = 0; i < (int)strlen(packetString); i++) {
       if (packetString[i] == '*') {
          seen++;
          if (seen == 2) {
